TVWeb/requirements.txt
ruamel.yaml==0.17.9
psycopg2-binary
flask-socketio
gevent
gevent-websocket
WTForms
flask-wtf
flask-sqlalchemy
Bootstrap-Flask
sqlacodegen
flask-sqlacodegen
flask-cors
requests
email_validator
Flask-Session
gunicorn
numpy
metadata_parser
scikit-learn
pandas
kneed
configparser
Flask-Mail
pyjwt
lxml
